Update README.md
-	In “ Set up the Smart Contract”:,
a)	In the modified code, require statements are added to perform error handling. The `getLeaderboard` and `updateLeaderboard` functions now verify that the provided index is within the valid range of existing leaderboards before accessing or modifying the data. If an invalid index is provided, an error message is returned to prevent unintended behaviour and ensure the contract's integrity.

-	Added relevant comments to the code as they provide a brief overview of what the contract does and explain the purpose of each function and variable.
-	Added an appropriate title ie- “Getting Started with Celo Blockchain using Rust Programming Language:” to the tutorial.
-	Added a proper conclusion.
-	Applied code indentation so that it looks neat and aligned.
-	Added relevant links to the source.
-	Did paraphrasing.
-	Added punctuations wherever, required.
-	Corrected spelling mistakes and grammatical errors.
-	Formatted the entire document by adding/removing ample spaces between paragraphs so that it looks well-structured.